Changed all Floats to Doubles. Implemented IBUs. Implemented DataHolder for Data transfer to other activities.

+++ app/src/main/java/com/example/cs121/final_project/Add_Ing_Activities_Dialogs/PickGrainDialog.java
@@ -131,13 +131,13 @@ public class PickGrainDialog extends Dialog implements
         }
         
         int timeparse = Integer.parseInt(time.getText().toString());
-        Float weightlbparse = Float.parseFloat(weight_lb.getText().toString()) * 16;
-        Float weightoz = Float.parseFloat(weight_oz.getText().toString());
+        Double weightlbparse = Double.parseDouble(weight_lb.getText().toString()) * 16;
+        Double weightoz = Double.parseDouble(weight_oz.getText().toString());
         if (time_type.getSelectedItem().toString().equals("Days")) timeparse *= 1440;
         Item grain = new Item(1, timeparse, name.getText().toString(),
                 type.getText().toString(), null, null, use.getSelectedItem().toString(),
-                Float.parseFloat(color.getText().toString()),
-                Float.parseFloat(potential.getText().toString()),
+                Double.parseDouble(color.getText().toString()),
+                Double.parseDouble(potential.getText().toString()),
                 (weightlbparse+weightoz), null, null);
         MyDialogFragmentListener activity = (MyDialogFragmentListener) c;
         activity.setItem(grain);
